🩻 Added Loading Skeleton
We've created a loading skeleton to let the user know everything is loading.

Also, we redeveloped the player initialization to make sure everything is loaded before the user can interact with something.

This is because as a PWA, this can be installed as an app. If they have no indication something is loading, they will try to launch something that isn't ready yet.