💾 Settings est arrive
We've added some settings so that the volume slider is saved in the browser. Handy for if you want to return later and that you're headphones aren't blasted of your head!

Also, we've fixed some issue with the tags. It seems that we've forgot to develop a method to reload them every amount of time. This is now fixed.